requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
psycopg2-binary
python-dotenv
pillow
django-filter
docker
pytest
pytest-django
pytest-cov
stripe
celery
django-celery-beat
redis

flake8
ipython
black
isort